Refactor : Game system 리팩토링
상세 :
- 의미 단위로 모듈 통합
- 가독성을 위해 require 구조분해할당 제거
- gameController, roomController, playController로 통합

#315
