--- server/socket/game/assignStreamer.js
@@ -1,7 +1,0 @@
-const io = require('../io');
-
-const assignStreamer = streamer => {
-  io.to(streamer.socketId).emit('assignStreamer');
-};
-
-module.exports = assignStreamer;
