--- server/socket/game/assignViewer.js
@@ -1,9 +1,0 @@
-const io = require('../io');
-
-const assignViewer = (viewer, streamer) => {
-  io.to(viewer.socketId).emit('assignViewer', {
-    streamerSocketId: streamer.socketId,
-  });
-};
-
-module.exports = assignViewer;
