--- server/socket/game/index.js
@@ -1,17 +1,0 @@
-const resetGameProgress = require('./resetGameProgress');
-const prepareRound = require('./prepareRound');
-const prepareSet = require('./prepareSet');
-const assignStreamer = require('./assignStreamer');
-const assignViewer = require('./assignViewer');
-const getViewers = require('./getViewers');
-const getNextStep = require('./getNextStep');
-
-module.exports = {
-  resetGameProgress,
-  prepareRound,
-  prepareSet,
-  assignStreamer,
-  assignViewer,
-  getViewers,
-  getNextStep,
-};
