+++ server/socket/index.js
@@ -7,6 +7,9 @@ const {
 const io = require('./io');
 
 io.on('connection', socket => {
+  /**
+   * 게임 전
+   */
   socket.on('askSocketId', () => {
     socket.emit('sendSocketId', { socketId: socket.id });
   });
@@ -14,11 +17,22 @@ io.on('connection', socket => {
 
   socket.on('sendReady', sendReadyHandler.bind(null, socket));
 
+  /**
+   * 게임로직 필요
+   * ....
+   */
+
+  /**
+   * 채팅
+   */
   socket.on(
     'sendChattingMessage',
     sendChattingMessageHandler.bind(null, socket),
   );
 
+  /**
+   * socket disconnect 관리
+   */
   socket.on('disconnecting', disconnectingHandler.bind(null, socket));
 
   socket.on('sendDescription', ({ target, description }) => {
