--- server/socket/player/index.js
@@ -1,5 +1,0 @@
-const createPlayer = require('./createPlayer');
-const setPlayer = require('./setPlayer');
-const getPlayerBySocket = require('./getPlayerBySocket');
-
-module.exports = { createPlayer, setPlayer, getPlayerBySocket };
