--- server/socket/room/getRoomByRoomId.js
@@ -1,8 +1,0 @@
-const rooms = require('../rooms');
-
-const getRoomByRoomId = roomId => {
-  const room = rooms[roomId];
-  return room;
-};
-
-module.exports = getRoomByRoomId;
