--- server/socket/room/index.js
@@ -1,17 +1,0 @@
-const generateRoomId = require('./generateRoomId');
-const getJoinableRoomId = require('./getJoinableRoomId');
-const initializeRoom = require('./initializeRoom');
-const joinRoom = require('./joinRoom');
-const getRoomByRoomId = require('./getRoomByRoomId');
-const setRoom = require('./setRoom');
-const removePlayerFromRoom = require('./removePlayerFromRoom');
-
-module.exports = {
-  generateRoomId,
-  getJoinableRoomId,
-  initializeRoom,
-  joinRoom,
-  getRoomByRoomId,
-  setRoom,
-  removePlayerFromRoom,
-};
