--- server/socket/room/joinRoom.js
@@ -1,9 +1,0 @@
-const rooms = require('../rooms');
-
-const joinRoom = ({ socket, roomId, player }) => {
-  socket.join(roomId);
-  socket.roomId = roomId;
-  rooms[roomId].players[player.socketId] = player;
-};
-
-module.exports = joinRoom;
